(closes #19 - feature/code-gen-framework/19)
Squashed commit of the following:

commit 2d8a0f216ce4b715e60a7f24634979f293382de9
Author: Brandon Li <brandon.li820@icloud.com>
Date:   Tue Jan 17 16:27:42 2023 -0800

    consistent test

commit af8bce0dbde8c1da5efa7b25ca141442d1e4bc6a
Author: Brandon Li <brandon.li820@icloud.com>
Date:   Tue Jan 17 16:25:01 2023 -0800

    cleanup

commit d8c7fd35ba3aa37e85130f89ac39af808e1d3e11
Author: Brandon Li <brandon.li820@icloud.com>
Date:   Tue Jan 17 16:22:57 2023 -0800

    clean up boostrapper

commit 0aa60b2a130df14fa4a45222cfa9c165865f1a49
Author: Brandon Li <brandon.li820@icloud.com>
Date:   Tue Jan 17 15:33:42 2023 -0800

    print numbers

commit 4e51807941375290086fa3ee4406ae53743175f3
Merge: 57013d5 fafba1f
Author: Brandon Li <brandon.li820@icloud.com>
Date:   Tue Jan 17 12:51:25 2023 -0800

    merge

commit 57013d57d63894c5d53e7c03266f51c3699d8709
Author: Brandon Li <brandon.li820@icloud.com>
Date:   Sun Jan 15 15:14:35 2023 -0800

    adjust

commit a1ace57d91cdd01ee005888fa9e3d4b43752a29b
Author: Brandon Li <brandon.li820@icloud.com>
Date:   Sun Jan 15 00:23:35 2023 -0800

    bootstrap

commit 971fd04201e246e8954aa2fadbbf3d8ce2c18abe
Author: Brandon Li <brandon.li820@icloud.com>
Date:   Sun Jan 15 00:21:33 2023 -0800

    bootstrap

+++ src/asm/asm.rs
@@ -24,7 +24,7 @@ pub enum Register {
 /// Operands for different instructions.
 pub enum Operand {
     Reg(Register),
-    Imm(i32),
+    Imm(i64),
     MemOffset(Box<Operand>, Box<Operand>),
 }
